Redirect user to the dashboard, if there are no non-fixed events
